+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Selection of active panel when `active` is provided to main navigation items.
+
 ## [3.2.0] - 2024-01-24
 
 ### Added